requirements.txt
customtkinter
yt-dlp
ffmpeg
Pillow
